+++ Projects/Feature/Ledger/Interface/Service/LedgerServiceInterface.swift
@@ -1,0 +1,7 @@
+/// Ledger에서 사용하는 전역서비스 모음
+public protocol LedgerServiceInterface {
+  var agency: AgencyServiceInterface { get }
+  var member: MemberServiceInterface { get }
+  var ledgerList: LedgerListServiceInterface { get }
+}
+
