+++ Projects/Feature/Ledger/Sources/Scene/Ledger/TabItems/Member/MemberList/MemberTabReactor.swift
@@ -5,6 +5,7 @@ import BaseDomain
 import DesignSystem
 import UserInterface
 import Utility
+import LedgerFeatureInterface
 
 final class MemberTabReactor: Reactor {
