+++ Projects/Feature/Ledger/Sources/Scene/Service/LedgerListService.swift
@@ -1,19 +1,7 @@
 import BaseDomain
+import LedgerFeatureInterface
 
 import RxSwift
-
-public enum LedgerListEvent {
-  case selectedDateRange(start: DateInfo, end: DateInfo)
-  case createLedgerRecord
-  case update
-}
-
-public protocol LedgerListServiceInterface {
-  var event: PublishSubject<LedgerListEvent> { get }
-  func selectedDate(start: DateInfo, end: DateInfo) -> Observable<Void>
-  func createLedgerRecord() -> Observable<Void>
-  func updateList()
-}
 
 final class LedgerListService: LedgerListServiceInterface {
   let event = PublishSubject<LedgerListEvent>()
