+++ Projects/Feature/Ledger/Sources/Scene/Service/LedgerService.swift
@@ -1,9 +1,4 @@
-/// Ledger에서 사용하는 전역서비스 모음
-public protocol LedgerServiceInterface {
-  var agency: AgencyServiceInterface { get }
-  var member: MemberServiceInterface { get }
-  var ledgerList: LedgerListServiceInterface { get }
-}
+import LedgerFeatureInterface
 
 public final class LedgerService: LedgerServiceInterface {
   public let agency: AgencyServiceInterface = AgencyService()
